+++ app/src/main/java/com/example/magickfinaljesus/Eventos.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 data class Eventos(
     val id:String?=null,
     val nombre:String?=null,
-    val precio:Int?=null,
+    val precio:Float?=null,
     val aforo_max:Int?=null,
     val aforo_ocupado:Int?=null,
     val fecha:String?=null,
